// frontend/src/types/enums/ColorClass.ts
/**
 * THIS FILE IS AUTOMATICALLY GENERATED BY genColors.js.
 *
 * DO NOT MANUALLY MODIFY THIS FILE.
 */

enum ColorClass {
  BackgroundOverlay = "colorBackgroundOverlay",
  BeigeDark = "colorBeigeDark",
  BeigeLight = "colorBeigeLight",
  Dimmer = "colorDimmer",
  Ghost = "colorGhost",
  Navy = "colorNavy",
  NavyHover = "colorNavyHover",
  Pink = "colorPink",
  Rust = "colorRust",
  WinterGreen = "colorWinterGreen",
  WinterGreenHover = "colorWinterGreenHover",
  Yellow = "colorYellow",
  YellowHover = "colorYellowHover",
}

export default ColorClass;